requirements.txt
python-telegram-bot==21.6
openai==1.55.0
httpx==0.27.2
